backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
pydantic
sqlalchemy
sqlite-lib
cors-middleware
